Allow escaping characters, multiple quote types in CommandString
It is now possible to use either " or ' quotes in a string, as well as opposite quote types in quoted substrings - eg. " ' " and ' " ' - and have the output appear as " ' " and " " ". Additionally, \ is now recognized as an escape character (although it is not validated, and escape sequences such as \n are not recognized) for quotes and back slashes - so " \" " appears as " " ", and "\\" appears as "\".