Add free tags to books, change ExtendedTag to NamedTag
Books might have both named tags - additional metadata which should store specific information, such as modification dates, but is not included in the program by default, and free tags - or general tags which vaguely describe a book. This commit distinguishes the two by adding NamedTag and FreeTag, respectively.

It also refactors specific SQLite field names - tag_name and tag_value are shorted to name and value respectively, since tag_ is redundant on a TagData object.

Editing named tags is the same, but editing free tags is done with `:e tag tag_name`.

Additionally, the ability to search for tags is now present, using the exclusive keyword `tag`. These are searched on a tag by tag basis, and can not be used as a reference for sorting, due to random orderings.