Make application async
This involves entirely removing Arc<RwLock<Book>>, because this seems like a terrible idea, and lets us invalidate the cache by mutating books - inarguably, a terrible choice. However, this currently comes at the cost of excessive copying - which I will fix at a future date, by maing the books Arc<>.

Otherwise, we have fixed a lot of particularly terrible decisions related to handling these books in an async fashion.

In the future, it seems that removing any external copies of Book, and instead storing BookId and fetching the book when needed is probably the best choice.