Rename bookmap.rs to cache.rs, limit capacity by default
Instead of always keeping every book in memory, only a random subset of recently used books will be kept - this is currently implemented as a simple removal of the next key in the hashmap. This isn't ideal, but at the same, I didn't find any particular nice LRU caches, so it'll have to do.